+++ CHANGELOG.md
@@ -5,6 +5,9 @@
 The following components have had minor internal changes to satisfy the introduction of stricter linting rules:
 
 * AppWrapper
+* Dropdown
+* DropdownFilter
+* DropdownFilterAjax
 * Carousel
 * Checkbox
 * Column
@@ -38,6 +41,7 @@ The following components have had minor internal changes to satisfy the introduc
 * `Menu` has been updated to use a `<nav>` tag as its root element.
 * `InlineInputs` children are now wrapped by Columns by the component
 * `TableHeader`: improve accessibility of sortable columns. They can now receive focus via the keyboard, and include `aria-sort` and `aria-label` attributes to indicate they are sortable, the current sort direction, and which direction the column will be sorted when sorting is next activated.
+* `Browser`: add a new method `setInputFocus` to focus on the input field of passed in ref but does not select text
 
 ## Deployment Changes